+++ lib/enums/selected_item_align.dart
@@ -1,0 +1,5 @@
+enum SelectedItemAlign {
+  start,
+  center,
+  end,
+}