+++ lib/widgets/dropdown_widget.dart
@@ -185,6 +185,33 @@ class DropdownWidgetState<T> extends State<DropdownWidget<T>> {
               ),
             ),
           ),
+
+          /// center line test
+          // Positioned(
+          //   top: dropdownOffset.dy - widget.dropdownOptions.marginGap.top,
+          //   left: dropdownOffset.dx - widget.dropdownOptions.marginGap.left,
+          //   child: IgnorePointer(
+          //     child: GestureDetector(
+          //       onTap: () {},
+          //       child: _buildAnimation(
+          //         child: Container(
+          //           margin: widget.dropdownOptions.marginGap,
+          //           width: _dropdownCalculator.dropdownWidth,
+          //           height: _dropdownCalculator.dropdownHeight +
+          //               widget.dropdownOptions.borderSide.width,
+          //           padding: EdgeInsets.all(
+          //               widget.dropdownOptions.borderSide.width * 0.5),
+          //           child: Center(
+          //             child: Container(
+          //               height: 3,
+          //               color: Colors.blue.withOpacity(0.8),
+          //             ),
+          //           ),
+          //         ),
+          //       ),
+          //     ),
+          //   ),
+          // )
         ],
       ),
     );